% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familias2ped.R
\name{connectedComponents}
\alias{connectedComponents}
\title{Connected pedigree components}
\usage{
connectedComponents(id, fid, mid)
}
\arguments{
\item{id}{A vector of ID labels (character or numeric)}

\item{fid}{The ID labels of the fathers (or "0" if missing)}

\item{mid}{The ID labels of the mothers (or "0" if missing)}
}
\value{
A list, where each element is a subset of \code{id} constituting a
connected pedigree
}
\description{
Compute the connected parts of a (possibly) disconnected pedigree. This is a
necessary step when converting pedigree data from software like Familias
(which allows disconnected pedigrees) to \code{pedtools} (which requires pedigrees
to be connected).
}
