% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familias-deprecated.R
\name{familias}
\alias{familias}
\alias{readFam}
\title{Import/export from Familias}
\usage{
readFam(...)
}
\arguments{
\item{...}{Arguments passed on to the respective \code{pedFamilias} function.}
}
\description{
Functions for reading .fam files associated with the Familias
software for forensic kinship computations.

\emph{\strong{Deprecated}} These functions have been moved to a separate package,
\code{pedFamilias}, and will be removed from \code{forrel} in a future version.
}
