% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familias2ped.R
\name{Familias2ped}
\alias{Familias2ped}
\alias{readFamiliasLoci}
\title{Convert \code{Familias} objects to \code{ped} format}
\usage{
Familias2ped(familiasped, datamatrix, loci)

readFamiliasLoci(loci)
}
\arguments{
\item{familiasped}{A \code{\link[Familias:FamiliasPedigree]{Familias::FamiliasPedigree()}} object or a list of such.}

\item{datamatrix}{A data frame with two columns per marker (one for each
allele) and one row per individual.}

\item{loci}{A \code{\link[Familias:FamiliasLocus]{Familias::FamiliasLocus()}} object or a list of such.}
}
\value{
A \code{ped} object, or a list of such.
}
\description{
Familias is a widely used program for computations in forensic genetics. The
function documented here converts pedigrees and marker data from Familias to
\code{\link[pedtools:ped]{pedtools::ped()}} format, thus enabling such data to be analysed with
\code{forrel}. This may be of interest for specialized computations not
implemented in Familias, e.g. conditional simulations.
}
\details{
The Familias program represents pedigrees and marker data in a way that
differs from \code{pedtools} in several ways, mostly because of the latter's
stricter definition of a \emph{pedigree}. In \code{pedtools} pedigrees must be
connected, and each member must have either 0 or 2 parents present in the
pedigree. None of this is required by \code{FamiliasPedigree} objects. The
conversion function \code{Familias2ped} takes care of all potential differences:
It converts each FamiliasPedigree into a list of connected ped objects,
adding missing parents where needed.
}
\references{
Windows Familias is freely available from \url{http://familias.name}.
}
\author{
Magnus Dehli Vigeland, Thore Egeland
}
