% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEstimate.R
\name{fEstimate}
\alias{fEstimate}
\title{Estimation of inbreeding coefficients}
\usage{
fEstimate(
  x,
  ids = typedMembers(x),
  method = c("mle", "simple", "ritland"),
  markers = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{ped} object or a list of such.}

\item{ids}{A vector with ID labels, by default, all genotyped members of \code{x}.}

\item{method}{A character, either "mle" (maximum-likelihood estimate;
default), "simple" (simple method-of-moments estimator), or "ritland" (see
Details).}

\item{markers}{A vector with names or indices of markers attached to x,
indicating which markers to include. By default, all markers are included.
Missing genotypes are removed before estimation.}

\item{verbose}{A logical.}
}
\value{
A named numeric vector.
}
\description{
Estimate the inbreeding coefficient \eqn{f} of pedigree members from their
genotypes. The default method uses maximum likelihood, but moment estimators
are also available.
}
\details{
The ML estimator assumes independent markers and known allele frequencies.
For a single marker with allele frequencies \eqn{p_a}, the genotype
likelihood \eqn{P(G \mid f)} is:

\deqn{P(G = a/a \mid f) = f p_a + (1-f)p_a^2,} \deqn{P(G = a/b \mid f) =
2(1-f)p_a p_b,\quad a \neq b.}

For multiple markers, the full likelihood is the product over markers. The
function uses \code{optimise()} to maximise the log-likelihood over \eqn{[0,1]}.

The \code{"simple"} method is a method-of-moments estimator based on observed and
expected homozygosity. Let \eqn{L} be the number of markers, \eqn{N_H} the
number of observed homozygous genotypes, and \eqn{H^* = \sum_1^L \sum_a
p_a^2} the expected number under Hardy--Weinberg equilibrium. The estimator
is then:

\deqn{\hat{f} = \dfrac{N_H - H^*}{L - H^*}.}

The \code{"ritland"} method implements the estimator of Ritland (1996). For marker
\eqn{j}, let \eqn{n_j} be the number of alleles, and define \eqn{T_j} to be
\eqn{1/p_a} if the genotype is \eqn{a/a}, and 0 otherwise. The estimator is
then:

\deqn{\hat{f} = \dfrac{\sum_{j=1}^L (T_j - 1)}{\sum_{j=1}^L (n_j - 1)}
      = \dfrac{\sum_{\text{hom}} 1/p_a - L}{A - L},}

where \eqn{A = \sum n_j} is the total number of alleles across markers. Note
that both this and the "simple" estimator may produce estimates outside the
interval \eqn{[0,1]}.

The MLE approach is described explicitly in Section 3.2 of Thompson (1986).
Ritland (1996) introduces the Ritland estimator, and also mentions both the
simple estimator and the MLE. Note that Ritlands formulation of the MLE is
slightly different (but equivalent) to Thompson's.
}
\examples{

# Simulate DNA profile for child of full sibs
x = nuclearPed(2, sex = 1:2) |> addSon(3:4) |>
  profileSim(ids = 5, markers = NorwegianFrequencies, seed = 123)
x

# Estimate inbreeding coefficient (expected: 0.0625)
fEstimate(x)

#-----------------------------
# Compare different estimators
#-----------------------------

# Simulate 200 DNA profiles for a child of half-sibs
x = halfSibPed(sex2 = 2) |> addSon(4:5) |>
  profileSim(N = 200, ids = 6, markers = NorwegianFrequencies, seed = 123)

# Estimates using all three methods
fhat = data.frame(
  mle = sapply(x, fEstimate, method = "mle"),
  sim = sapply(x, fEstimate, method = "simple"),
  rit = sapply(x, fEstimate, method = "ritland")
)

# Mean estimates
apply(fhat, 2, mean)

# RMSE
apply(fhat, 2, function(v) sqrt(mean((v - 0.125)^2)))

}
\references{
\itemize{
\item Thompson, E. A. (1986). Pedigree Analysis in Human Genetics.
\emph{Johns Hopkins University Press}.
\item Ritland, K. (1996). Estimators for pairwise relatedness and individual
inbreeding coefficients. \emph{Genetical Research}.
}
}
\seealso{
\code{\link[=ibdEstimate]{ibdEstimate()}}
}
