% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FORCE}
\alias{FORCE}
\title{FORCE panel kinship SNPs}
\format{
A data frame with 3915 rows and 6 columns:
\itemize{
\item MARKER: Marker name (rs number)
\item CHROM: Chromosome (1-22)
\item MB: Physical position in megabases, according to genome build GRCh38
\item AL1: Major allele
\item AL2: Minor allele
\item FR1: Allele frequency of AL1
}

The data frame is designed to work smoothly with \code{\link[pedtools:setSNPs]{pedtools::setSNPs()}}.
}
\source{
Tillmar et al. The FORCE Panel: An All-in-One SNP Marker Set for
Confirming Investigative Genetic Genealogy Leads and for General Forensic
Applications. Genes. (2021)
}
\usage{
FORCE
}
\description{
A data frame with data on a subset of the SNPs included on the FORCE panel
(Tillmar et al., 2021). The subset contains the SNPs marked recommended for
kinship analysis. As the original publication did not include allele
frequencies, these have been obtained from Ensemble via the biomaRt package.
15 markers were removed as frequency information could not be
retrieved.
}
\examples{
x = setSNPs(nuclearPed(), FORCE)
summary(x)
head(getMap(x))

}
\keyword{datasets}
