% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forrel-package.R
\docType{package}
\name{forrel}
\alias{forrel}
\alias{forrel-package}
\title{forrel: Relatedness inference and forensic pedigree analysis}
\description{
Forensic applications of pedigree analysis, including marker simulation,
relatedness inference and exclusion probailities. Data import from the
Familias software is fully supported.
}
