% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{Likelihood ratios of pedigree hypotheses}
\usage{
LR(x, ref, markers)
}
\arguments{
\item{x}{A list of pedigree alternatives. Each alternative should be either a
single \code{ped} object or a list of such.}

\item{ref}{A single integer, indicating the index (in the list \code{x}) of the
reference alternative. This is used in the denominator of each LR.}

\item{markers}{A vector of integers, indexing which markers should be
included. If NULL (the default) all markers are used.}
}
\value{
A list with entries
\itemize{
\item \code{LR} : Likelihood ratios
\item \code{LRperMarker} : Likelihood ratios for each marker
\item \code{likelihoodsPerSystem} : Likelihoods for each marker
\item \code{time} user system and elapsed time
}
}
\description{
This function computes likelihood ratios for a given a list of pedigrees
(\code{ped} and/or \code{singleton} objects), with genotype data from the same set of
markers. Data exported from the 'Familias' software can be analysed by using
\code{\link[=Familias2ped]{Familias2ped()}} prior to calling this function.
}
\examples{

# Simulate genotypes for 5 tetra-allelic markers for a pair of full sibs
set.seed(123)
sibs = nuclearPed(children = c("A", "B"))
sibs = simpleSim(sibs, N = 5, alleles = 1:4, ids = c("A", "B"))

# Create two alternative hypotheses and transfer the simulated genotypes to them
halfsibs = relabel(halfSibPed(), old = 4:5, new = c("A", "B"))
halfsibs = transferMarkers(sibs, halfsibs)

unrel = list(singleton("A"), singleton("B"))
unrel = transferMarkers(sibs, unrel)

# Compute LR with 'unrelated' as reference
LR(list(sibs, halfsibs, unrel), ref=3)

}
\author{
Magnus Dehli Vigeland and Thore Egeland
}
