% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonIP.R
\name{missingPersonIP}
\alias{missingPersonIP}
\title{LR simulation for missing person cases}
\usage{
missingPersonIP(
  reference,
  missing,
  markers = NULL,
  nsim = 1,
  threshold = NULL,
  disableMutations = NA,
  seed = NULL,
  true = c("missing", "unrelated"),
  verbose = TRUE
)
}
\arguments{
\item{reference}{A \code{ped} object with attached markers.}

\item{missing}{The ID label of the missing pedigree member.}

\item{markers}{Names or indices of attached markers to include. By default all markers.}

\item{nsim}{A positive integer: the number of simulations}

\item{threshold}{A numeric vector with one or more positive LR thresholds.}

\item{disableMutations}{This parameter determines how mutation models are
treated. Possible values are as follows:
\itemize{
\item \code{NA} (the default): Mutations are disabled only for those markers whose
known genotypes are consistent with the pedigree. This is determined by
temporarily removing all mutation models and checking which markers have
nonzero likelihood.
\item \code{TRUE}: Mutations are disabled for all markers. This will result in an
error if any markers are inconsistent.
\item \code{FALSE}: No action is done to disable mutations.
\item A vector containing the names or indices of those markers for which
mutations should be disabled.
}}

\item{seed}{An integer seed for the random number generator (optional).}

\item{true}{Either "missing" (default) or "unrelated", indicating whether profiles are
simulated for the true missing person or for an unrelated person.}

\item{verbose}{A logical.}
}
\value{
A \code{mpIP} object, which is essentially a list with the following entries:
\itemize{
\item \code{LRperSim}: A numeric vector of length \code{nsim} containing the total LR for each
simulation.
\item \code{log10LRperSim}: The corresponding total LRs on the log10 scale.
\item \code{meanLRperMarker}: The mean LR per marker, over all simulations.
\item \code{meanLR}: The mean total LR over all simulations.
\item \code{meanLogLR}: The mean of \code{log10LRperSim}.
\item \code{IP}: A named numeric of the same length as \code{threshold}. For each element of
\code{threshold}, the fraction of simulations resulting in an LR exceeding the given
number. With \code{true = "missing"} this is the inclusion power; with \code{true =   "unrelated"} it is the corresponding fraction among unrelated persons.
\item \code{params}: A list containing the input parameters \code{missing}, \code{markers}, \code{nsim},
\code{threshold}, \code{seed}, \code{disableMutations} and \code{true}.
}
}
\description{
This function simulates the LR distribution in a missing person case, either for the
true missing person or for an unrelated person. The output contains the total LR and
log10 LR of each simulation, and various summary statistics. If a specific LR threshold
is given, the fraction of simulations exceeding the threshold is computed. When
simulating the true missing person (the default), this fraction is referred to as the
\emph{inclusion power} (Vigeland et al., 2020).
}
\examples{

# Two brothers are looking for their missing sibling.
# They are typed with 5 triallelic markers.
x = nuclearPed(3) |>
  markerSim(N = 5, ids = 3:4, alleles = 1:3, seed = 123, verbose = FALSE)

missingPersonPlot(x, missing = 5)

nsim = 20 # increase!

# Inclusion power statistics
ip = missingPersonIP(x, missing = 5, nsim = nsim, threshold = c(10, 100))
ip
head(ip$LRperSim)

# Simulate LRs for a random unrelated person
ip2 = missingPersonIP(x, missing = 5, nsim = nsim, threshold = c(10, 100),
                      true = "unrelated")
ip2
head(ip2$LRperSim)

# Plot distributions
LRpowerPlot(data = list(ip, ip2), threshold = 100)

}
\references{
Vigeland MD, Marsico FL, Herrera Piñero M, Egeland T (2020). "Prioritising
family members for genotyping in missing person cases: A general approach combining
the statistical power of exclusion and inclusion."
\emph{FSI: Genetics}, 49, 102376. \doi{10.1016/j.fsigen.2020.102376}
}
\seealso{
\code{\link[=missingPersonEP]{missingPersonEP()}}, \code{\link[=missingPersonLR]{missingPersonLR()}}, \code{\link[=missingPersonPlot]{missingPersonPlot()}}
}
