% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NorwegianFrequencies}
\alias{NorwegianFrequencies}
\title{Norwegian STR frequencies}
\format{
A list of length 35. Each entry is a numerical vector summing to 1,
named with allele labels.

The following markers are included:
\itemize{
\item \code{D3S1358}: 12 alleles
\item \code{TH01}: 10 alleles
\item \code{D21S11}: 26 alleles
\item \code{D18S51 }: 23 alleles
\item \code{PENTA_E}: 21 alleles
\item \code{D5S818}: 9 alleles
\item \code{D13S317}: 9 alleles
\item \code{D7S820}: 19 alleles
\item \code{D16S539}: 9 alleles
\item \code{CSF1PO}: 11 alleles
\item \code{PENTA_D}: 24 alleles
\item \code{VWA}: 12 alleles
\item \code{D8S1179}: 12 alleles
\item \code{TPOX}: 9 alleles
\item \code{FGA}: 25 alleles
\item \code{D19S433}: 17 alleles
\item \code{D2S1338}: 13 alleles
\item \code{D10S1248}: 9 alleles
\item \code{D1S1656}: 17 alleles
\item \code{D22S1045}: 9 alleles
\item \code{D2S441}: 13 alleles
\item \code{D12S391}: 23 alleles
\item \code{SE33}: 55 alleles
\item \code{D7S1517}: 11 alleles
\item \code{D3S1744}: 8 alleles
\item \code{D2S1360}: 10 alleles
\item \code{D6S474}: 6 alleles
\item \code{D4S2366}: 7 alleles
\item \code{D8S1132}: 12 alleles
\item \code{D5S2500}: 8 alleles
\item \code{D21S2055}: 18 alleles
\item \code{D10S2325}: 10 alleles
\item \code{D17S906}: 78 alleles
\item \code{APOAI1}: 41 alleles
\item \code{D11S554}: 51 alleles
}
}
\source{
Dupuy et al. (2013): \emph{Frequency data for 35 autosomal STR markers in
a Norwegian, an East African, an East Asian and Middle Asian population and
simulation of adequate database size}. Forensic Science International:
Genetics Supplement Series, Volume 4 (1).
}
\usage{
NorwegianFrequencies
}
\description{
A database of Norwegian allele frequencies for 35 STR markers.
}
\keyword{datasets}
