% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEstimate.R
\name{parentChildLikelihood}
\alias{parentChildLikelihood}
\title{Parent-child pedigree likelihoods}
\usage{
parentChildLikelihood(x, parent = NULL, child = NULL, f = NULL, check = TRUE)
}
\arguments{
\item{x}{A \code{ped} object with exactly two typed members.}

\item{parent}{ID of parent; identified automatically if missing.}

\item{child}{ID of child; identified automatically if missing.}

\item{f}{Inbreeding coefficient of child, calculated with
\code{ribd::inbreeding()} if missing.}

\item{check}{A logical indicating whether to check parent and child in \code{x}.}
}
\value{
A numeric vector with per-marker likelihoods P(parent & child
genotypes | f).
}
\description{
Fast likelihood computation of pedigrees in which the only genotyped
individuals are a (noninbred) parent and their (possibly inbred) child. In
this case the pedigree information is fully captured by the inbreeding
coefficient \eqn{f} of the child, enabling efficient calculations by
vectorising over the markers.
}
\examples{

db = NorwegianFrequencies[1:10]

# Simulate genotypes for a child of first cousins, and his mother
x = cousinPed(1, child = TRUE) |> profileSim(ids = 8:9, markers = db)

plot(x, hatched = typedMembers)

liks = parentChildLikelihood(x)

# Compare with standard likelihood calculation, requiring loop breaking etc.
stopifnot(all.equal(liks, pedprobr::likelihood(x)))

# bench::mark(parentChildLikelihood(x), pedprobr::likelihood(x))

}
