% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPlot.R
\name{powerPlot}
\alias{powerPlot}
\title{Exclusion/inclusion power plots}
\usage{
powerPlot(
  ep,
  ip,
  type = 1,
  ellipse = FALSE,
  col = NULL,
  labs = NULL,
  alpha = 1,
  shape = 1,
  size = 2,
  hline = NULL,
  vline = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{ep}{Exclusion power data, typically in the form of one or several
\code{EPresult} objects (output from \code{\link[=missingPersonEP]{missingPersonEP()}}. See Details and
Examples.}

\item{ip}{Inclusion power data, typically in the form of one or several
\code{mpIP} objects (output from \code{\link[=missingPersonIP]{missingPersonIP()}}. See Details and Examples.
The list structure of \code{ip} must be identical to that of \code{ep}.}

\item{type}{Plot type; either 1, 2 or 3.}

\item{ellipse}{A logical. If TRUE, data ellipsis are drawn for each inner
list of \code{ep}/\code{ip} containing more than 1 element.}

\item{col}{A color vector, recycle to match the top level length of \code{ep}.}

\item{labs}{A character of the same length as \code{ep}. If NULL, the names of
\code{ep} are used, if present.}

\item{alpha}{Transparency for minor points (see Details).}

\item{shape, size}{Plotting parameters}

\item{hline, vline}{Single numbers indicating positions for
horizontal/vertical "threshold" lines. When \code{type = 1}, these determine the
shaded vertical regions, by default starting at 0.95.}

\item{xlim, ylim}{Axis limits; automatically chosen if NULL.}

\item{xlab, ylab}{Axis labels; automatically chosen if NULL.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
This function offers three different visualisations of exclusion/inclusion
powers, particularly for missing person cases.
}
\details{
The plot types are as follows:

\code{type = 1}: x = Exclusion power; y = Inclusion power

\code{type = 2}: x = Exclusion odds ratio; y = Inclusion odds ratio

\code{type = 3}: x = Expected number of exclusions; y = average log(LR)

For each \code{EPresult} object in \code{ep}, and the corresponding element of \code{ip},
the relevant data is extracted from each, producing a single point the final
plot.

In the most general case \code{ep} (and similarly for \code{ip}) can be a list of lists
of \code{EPresult} objects. To simplify the discussion we refer to the inner lists
as "groups". A group may consist of a single point, or several (typically
many simulations of the same situation). Points within the same group are
always drawn with the same color and shape.

When plotting several groups, two sets of points are drawn:
\itemize{
\item Major points: Group means.
\item Minor points: Individual points in groups with more than one element.
}
}
\examples{

ref = nuclearPed(father = "fa", mother = "mo", child = "MP")
ref = setMarkers(ref, marker(ref, alleles = 1:2))

# Alternatives for genotyping
sel = list("fa", c("fa", "mo"))

# Simulate power for each selection
simData = MPPsims(ref, selections = sel, nProfiles = 10,
                  lrSims = 10, thresh = 2)

# Extract EP and IP data
epSims = lapply(simData, '[[', 'ep')
ipSims = lapply(simData, '[[', 'ip')

# Power plot 1: EP vs IP
powerPlot(epSims, ipSims, type = 1)

# Power plot 3: Expected number of exclusions vs E[LR]
powerPlot(epSims, ipSims, type = 3, hline = log10(2), vline = 1)

}
\seealso{
\code{\link[=missingPersonEP]{missingPersonEP()}}, \code{\link[=missingPersonEP]{missingPersonEP()}}, \code{\link[=MPPsims]{MPPsims()}}
}
