% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSim.R
\name{profileSim}
\alias{profileSim}
\title{Simulation of complete DNA profiles}
\usage{
profileSim(
  x,
  N = 1,
  ids = NULL,
  markers = NULL,
  loopBreakers = NULL,
  seed = NULL,
  numCores = 1,
  simplify1 = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{ped} object or a list of such.}

\item{N}{The number of complete simulations to be performed.}

\item{ids}{A character (or coercible to character) of ID labels indicating
whose genotypes should be simulated. Alternatively, a function taking \code{x}
as input and returning a character vector of ID labels.}

\item{markers}{Either a vector indicating a subset of markers attached to
\code{x} (default: all), or a named list of frequency vectors. In the former
case, simulations are conditional on all locus attributes (allele
frequencies, mutation models, etc) and any existing genotypes in the
indicated markers. If \code{markers} is a list of frequency vectors, these
are attached as new markers to \code{x} (removing any existing markers) and
used for unconditional simulations.}

\item{loopBreakers}{A vector containing IDs of individuals to be used as
loop breakers. Can usually be left as NULL, in which case automatic loop
breaking is applied. See \code{\link[pedtools:breakLoops]{pedtools::breakLoops()}}.}

\item{seed}{An integer seed for the random number generator. Used only when
no mirai workers are set. For parallel simulations, set the seed in
\code{\link[mirai:daemons]{mirai::daemons()}}.}

\item{numCores}{Deprecated. Use \code{\link[mirai:daemons]{mirai::daemons()}} to set the number of
workers.}

\item{simplify1}{A logical, by default TRUE, removing the outer list layer
when \code{N = 1}. See Value.}

\item{verbose}{A logical, by default TRUE.}
}
\value{
A list of \code{N} objects similar to \code{x}, but with simulated genotypes.
Any previously attached markers are replaced by the simulated profiles. If
the indicated markers contained genotypes for some pedigree members, these
are still present in the simulated profiles.

If \code{N = 1} and \code{simplify1 = TRUE}, the outer list layer is removed, i.e.,
\code{profileSim(..., N = 1, simplify1 = T)} is equivalent to \code{profileSim(..., N = 1, simplify1 = F)[[1]]}. This is usually the desired object in
interactive use, and works well with piping.

When using \code{profileSim()} in other functions, it is recommended to add
\code{simplify1 = FALSE} to safeguard against issues with \code{N = 1}.
}
\description{
Simulation of DNA profiles for specified pedigree members. Some pedigree
members may already be genotyped; in that case the simulation is conditional
on these. The main work of this function is done by \code{\link[=markerSim]{markerSim()}}.
}
\details{
\subsection{Parallel profile simulation}{

Simulations can be distributed across several R processes using the
\strong{mirai} package. To use this, start mirai workers before calling
\code{profileSim()}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mirai::daemons(8, seed = 123)

x = nuclearPed()
db = NorwegianFrequencies
res = profileSim(x, N = 100, markers = db)

mirai::daemons(0)
}\if{html}{\out{</div>}}

The call to \code{mirai::daemons(8, seed = 123)} starts 8 persistent workers and
configures reproducible parallel RNG. These workers remain available for
later calls until stopped with \code{mirai::daemons(0)}.

If no mirai daemons are set, \code{profileSim()} runs normally, and the \code{seed}
argument works as before. When mirai daemons are used, set the seed in
\code{mirai::daemons(n, seed = ...)} and omit the \code{seed} argument to \code{profileSim()}.
}
}
\examples{
# Example pedigree with two brothers
x = nuclearPed(children = c("B1", "B2"))

### Simulate profiles using built-in freq database
profileSim(x, markers = NorwegianFrequencies[1:3])

### Conditioning on known genotypes for one brother

# Attach two SNP markers with genotypes for B1
y = x |>
  addMarker(B1 = "1/2", alleles = 1:2) |>
  addMarker(B1 = "1",   alleles = 1:2, chrom = "X")

# Simulate 2 profiles of B2 conditional on the above
profileSim(y, N = 2, ids = "B2", seed = 123)

}
