% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankProfiles.R
\name{rankProfiles}
\alias{rankProfiles}
\title{Identify and rank the most likely profiles of a pedigree member}
\usage{
rankProfiles(x, id, markers = NULL, maxPerMarker = Inf, verbose = FALSE)
}
\arguments{
\item{id}{Name of the individual to be predicted.}

\item{markers}{Names or indices of the markers to be included. Default: all.}

\item{maxPerMarker}{The number of top candidates to be considered per marker.
Default: all.}

\item{verbose}{A logical, by default TRUE.}

\item{A}{\code{ped} object with attached markers.}
}
\value{
A list with three components. The first, a data frame with profiles
ranked according to likelihood. Under a flat prior, the posterior equals
the likelihood. Then follows the most likely profile. Finally, the second
to most likely genotypes are given.
}
\description{
Markers are assumed independent. For each marker, the possible genotypes for
\code{id} are evaluated and ranked according to how probable they are. It may be
wise to try first with \code{maxPerMarker = 1} to limit computation time,
particularly if mutations are modelled.
}
\examples{

x = nuclearPed(2, father = "FA") |>
  addMarker(`3` = "1/1", `4` = "1/2", alleles = 1:2, name = "m1") |>
  addMarker(`3` = "1/1", `4` = "2/2", alleles = 1:2, name = "m2") |>
  addMarker(`3` = "2/2", `4` = "2/2", alleles = 1:2, name = "m3")

rankProfiles(x, "FA")
rankProfiles(x, "FA", maxPerMarker = 2)
rankProfiles(x, "FA", maxPerMarker = 1)

# Same example with mutations allowed
y = setMutmod(x, model = "equal", rate = 0.002)
rankProfiles(y, "FA")

}
