% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XFORCE}
\alias{XFORCE}
\title{FORCE panel kinship SNPs (X-chromosomal)}
\format{
A data frame with 246 rows and 6 columns:
\itemize{
\item \code{CHROM}: Chromosome (all equal to "X")
\item \code{MARKER}: Marker name (rs number)
\item \code{MB}: Physical position in megabases (build GRCh37)
\item \code{A1}: First allele
\item \code{A2}: Second allele
\item \code{FREQ1}: Allele frequency of \code{A1}
}
}
\source{
Tillmar et al. The FORCE Panel: An All-in-One SNP Marker Set for
Confirming Investigative Genetic Genealogy Leads and for General Forensic
Applications. Genes. (2021)
}
\usage{
XFORCE
}
\description{
A data frame describing the 246 X-chromosomal SNPs included on the FORCE
panel for forensic genetics (Tillmar et al., 2021). As the original
publication did not include allele frequencies, these were downloaded from
Ensembl via the biomaRt package. For 9 markers the frequencies were obtained
manually from dbSNP.
}
\details{
To attach the XFORCE markers to a pedigree, use \code{\link[pedtools:setSNPs]{pedtools::setSNPs()}} (see
Examples).
}
\examples{
x = setSNPs(nuclearPed(), snpData = XFORCE)
summary(x)

getMap(x, markers = 1:5)
getFreqDatabase(x, markers = 1:5)

}
\keyword{datasets}
