% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPairwise.R
\name{checkPairwise}
\alias{checkPairwise}
\title{Check pedigree for relationship errors}
\usage{
checkPairwise(x, plot = TRUE, labels = FALSE, LRthreshold = 1000)
}
\arguments{
\item{x}{A \code{ped} object or a list of such.}

\item{plot}{A logical (default: TRUE). If TRUE, a plot is produced, showing
the IBD estimates in the IBD triangle.}

\item{labels}{A logical (default: FALSE). If TRUE, labels are included in the
IBD triangle plot.}

\item{LRthreshold}{A positive number (default: 1000). Estimates whose LR
(compared to the pedigree) exceeds}
}
\value{
A data frame containing both the estimated and pedigree-based IBD
coefficients for each pair of typed individuals.
}
\description{
This function provides a convenient way to check for pedigree errors, given
the available marker data. The function calls \code{\link[=IBDestimate]{IBDestimate()}} to estimate IBD
coefficients for all pairs of typed pedigree members, and computes the
likelihood ratio (LR) comparing each estimate to the coefficients implied by
the pedigree. By default, the estimates are shown in a colour-coded plot
where unlikely relationships are easy to spot.
}
\examples{

### Example with realistic data
x = addSon(nuclearPed(nch = 2), parent = 4)
x = setMarkers(x, locus = NorwegianFrequencies)
x = profileSim(x, N = 1, numCores = 1, seed = 1729)[[1]]

checkPairwise(x)

### Create sample swap between 3 and 7
als = getAlleles(x)
als[c(2,5), ] = als[c(5,2), ]
y = setAlleles(x, alleles = als)

checkPairwise(y)

\donttest{
# Nice plot of pedigree + triangle
dev.new(height = 5, width = 8, noRStudioGD = TRUE)
layout(rbind(1:2), widths = 2:3)
plot(y, margins = c(4,2,4,2))
checkPairwise(y)
}

}
\seealso{
\code{\link[=IBDestimate]{IBDestimate()}}, \code{\link[=IBDtriangle]{IBDtriangle()}}, \code{\link[=showInTriangle]{showInTriangle()}}
}
\author{
Magnus Dehli Vigeland
}
