% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familias2ped.R
\name{connectedComponents}
\alias{connectedComponents}
\title{Connected pedigree components}
\usage{
connectedComponents(ID, FID, MID)
}
\arguments{
\item{ID}{A vector of ID labels (character or numeric)}

\item{FID}{The ID labels of the fathers (or \code{0} if missing)}

\item{MID}{The ID labels of the mothers (or \code{0} if missing)}
}
\value{
A list, where each element is a vector of ID labels constituting a
connected pedigree
}
\description{
Compute the connected parts of a (possibly) disconnected pedigree. This is a
necessary step when converting pedigree data from software like Familias
(which allows disconnected pedigrees) to \code{pedtools} (which requires pedigrees
to be connected).
}
