% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusionPower.R
\name{exclusionPower}
\alias{exclusionPower}
\title{Power of exclusion}
\usage{
exclusionPower(ped_claim, ped_true, ids, markerindex = NULL,
  alleles = NULL, afreq = NULL, known_genotypes = list(),
  Xchrom = FALSE, plot = TRUE, verbose = T)
}
\arguments{
\item{ped_claim}{a \code{ped} object, or a list of several such objects,
describing the claimed relationship. If a list, the sets of ID labels must
be disjoint, that is, all ID labels must be unique.}

\item{ped_true}{a \code{ped} object, or a list of several such objects, describing
the true relationship. ID labels must be consistent with \code{ped_claim}.}

\item{ids}{individuals available for genotyping.}

\item{markerindex}{NULL, or a single numeric indicating the index of a marker
of \code{ped_claim} from which \code{alleles}, \code{afreq} and \code{known_genotypes} will be
extracted.}

\item{alleles}{a numeric or character vector containing marker alleles names.
Ignored if \code{markerindex} is non-NULL.}

\item{afreq}{a numerical vector with allele frequencies. An error is given if
they don't sum to 1 (rounded to 3 decimals). Ignored if \code{markerindex} is
non-NULL.}

\item{known_genotypes}{list of triplets \code{(a, b, c)}, indicating that
individual \code{a} has genotype \code{b/c}. Must be NULL if \code{markerindex} is
non-NULL.}

\item{Xchrom}{a logical: Is the marker on the X chromosome? Ignored if
\code{markerindex} is non-NULL.}

\item{plot}{either a logical or the character 'plot_only', controlling if a
plot should be produced. If 'plot_only', a plot is drawn, but no further
computations are done.}

\item{verbose}{a logical.}
}
\value{
A single numeric value. If \code{plot = "plot_only"}, the function returns
NULL after producing the plot.
}
\description{
Computes the power (of a single marker) of excluding a claimed relationship,
given the true relationship.
}
\details{
This function computes the 'Power of exclusion', as defined and discussed in
(Egeland et al., 2014).
}
\examples{

############################################
### A standard case paternity case:
### Compute the power of exclusion when the claimed father is in fact
### unrelated to the child.
############################################

# Claim: Individual 1 is the father of indiv 3
claim = nuclearPed(nch = 1, sex = 2)

# Truth: 1 and 3 are unrelated
true = list(singleton(id = 1), singleton(id = 3, sex = 2))

# Indivs 1 and 3 are available for genotyping
ids = c(1, 3)

# An equifrequent SNP
als = 2
afreq = c(0.5, 0.5)

# The exclusion power assuming no known genotypes
PE1 = exclusionPower(claim, true, ids = ids, alleles = als, afreq = afreq)

# Exclusion power if the child is known to have genotype 1/1:
PE2 = exclusionPower(claim, true, ids = ids, alleles = als, afreq = afreq,
                     known_genotypes = list(c(3, 1, 1)))

# Exclusion power if the SNP is on the X chromosome
PE3 = exclusionPower(claim, true, ids = ids, alleles = als, afreq = afreq,
                     Xchrom = TRUE)

stopifnot(PE1 == 0.125, PE2 == 0.25, PE3 == 0.25)

############################################
### Example from Egeland et al. (2012):
### Two females claim to be mother and daughter. Below we compute the power of various
### markers to reject this claim if they in reality are sisters.
############################################

mother_daughter = nuclearPed(1, sex = 2)
sisters = relabel(nuclearPed(2, sex = c(2, 2)), c(101, 102, 2, 3))

# Equifrequent SNP:
PE1 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters, ids = c(2, 3),
                     alleles = 2)

# SNP with MAF = 0.1:
PE2 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters, ids = c(2, 3),
                     alleles = 2, afreq = c(0.9, 0.1))

# Equifrequent tetra-allelic marker:
PE3 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters, ids = c(2, 3),
                     alleles = 4)

# Tetra-allelic marker with one major allele:
PE4 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters, ids = c(2, 3),
                     alleles = 4, afreq = c(0.7, 0.1, 0.1, 0.1))

stopifnot(round(c(PE1,PE2,PE3,PE4), 5) == c(0.03125, 0.00405, 0.08203, 0.03090))

####### How does the power change if the true pedigree is inbred?
sisters_LOOP = addParents(sisters, 101, father = 201, mother = 202)
sisters_LOOP = addParents(sisters_LOOP, 102, father = 201, mother = 203)

# Equifrequent SNP:
PE5 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters_LOOP,
                     ids = c(2, 3), alleles = 2)

# SNP with MAF = 0.1:
PE6 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters_LOOP,
                     ids = c(2, 3), alleles = 2, afreq = c(0.9, 0.1))

stopifnot(round(c(PE5,PE6), 5) == c(0.03125, 0.00765))

\dontrun{
# Equifrequent tetra-allelic marker:
PE7 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters_LOOP,
                     ids = c(2, 3), alleles = 4)

# Tetra-allelic marker with one major allele:
PE8 = exclusionPower(ped_claim = mother_daughter, ped_true = sisters_LOOP,
                     ids = c(2, 3), alleles = 4, afreq = c(0.7, 0.1, 0.1, 0.1))

stopifnot(round(c(PE7,PE8), 5) == c(0.07617, 0.03457))
}

}
\references{
T. Egeland, N. Pinto and M.D. Vigeland, \emph{A general approach to
power calculation for relationship testing.} Forensic Science
International: Genetics 9 (2014): 186-190. doi:10.1016/j.fsigen.2013.05.001
}
\author{
Magnus Dehli Vigeland
}
