% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedLR.R
\name{expectedLR}
\alias{expectedLR}
\title{Expected likelihood ratio}
\usage{
expectedLR(numeratorPed, denominatorPed, truePed = numeratorPed, ids, marker)
}
\arguments{
\item{numeratorPed}{A \code{ped} object or a list of such.}

\item{denominatorPed}{A \code{ped} object or a list of such.}

\item{truePed}{A \code{ped} object or a list of such.}

\item{ids}{A vector of ID labels corresponding to untyped pedigree members. (These must
be members of all three input pedigrees).}

\item{marker}{The name or index of a marker attached to \code{numeratorPed}. Alternatively,
a standalone \code{marker} object compatible with \code{numeratorPed}.}
}
\value{
A positive number; the expected LR.
}
\description{
This function computes the expected LR for a single marker, in a kinship test comparing
two hypothesised relationships between a set of individuals. The true relationship may
differ from both hypotheses. Some individuals may already be genotyped, while others
are available for typing. The implementation uses \code{oneMarkerDistribution()} to find the
joint genotype distribution for the available individuals, conditional on the known
data, in each pedigree.
}
\examples{

#---------
# Curious example showing that ELR may decrease
# by typing additional reference individuals
#---------

# Numerator ped
num = nuclearPed(father = "fa", mother = "mo", child = "ch")

# Denominator ped: fa, mo, ch are unrelated
den = singletons(c("fa", "mo", "ch"), sex = c(1, 2, 1))

# Scenario 1: Only mother is typed; genotype 1/2
p = 0.9
m = marker(num, mo = "1/2", afreq = c("1" = p, "2" = 1-p))
expectedLR(num, den, ids = "ch", marker = m)

1/(8*p*(1-p)) + 1/2  # exact

# Scenario 2: Include father, with genotype 1/1
genotype(m, id = "fa") = "1/1"
expectedLR(num, den, ids = "ch", marker = m)

1/(8*p*(1-p)) + 1/(4*p^2)  # exact

}
