% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familias2ped.R
\name{Familias2ped}
\alias{Familias2ped}
\alias{readFamiliasLoci}
\title{Convert 'Familias' output to ped objects}
\usage{
Familias2ped(familiasped, datamatrix, loci)

readFamiliasLoci(loci)
}
\arguments{
\item{familiasped}{A \code{\link[Familias:FamiliasPedigree]{Familias::FamiliasPedigree()}} object or a list of such.}

\item{datamatrix}{A data frame with two columns per marker (one for each
allele) and one row per individual.}

\item{loci}{A \code{\link[Familias:FamiliasLocus]{Familias::FamiliasLocus()}} object or a list of such.}
}
\value{
A \code{ped} object, or a list of such.
}
\description{
Familias is a widely used program for computations in forensic genetics. The
function documented here converts pedigrees and marker data from Familias to
\code{pedtools} format, thus enabling such data to be analysed with the \code{forrel}
package. This may be of interest for specialized computations not implemented
in Familias, e.g. conditional simulations.
}
\details{
The Familias program represents pedigrees and marker data in a way that
differs from \code{pedtools} in several ways, mostly because of the latter's
stricter definition of a \emph{pedigree}. In \code{pedtools}, unlike Familias, a
pedigree must be connected, and each member must have either 0 or 2 parents
present in the pedigree. None of this is required by \code{FamiliasPedigree}
objects. The conversion function \code{Familias2ped} takes care of all potential
differences: It converts each FamiliasPedigree into a list of connected ped
objects, adding missing parents where needed.
}
\examples{

\dontrun{
library(Familias)
library(pedtools)
data(NorwegianFrequencies)
TH01 = NorwegianFrequencies$TH01
locus1 = FamiliasLocus(TH01)
persons = c('mother', 'daughter', 'AF')
ped1 = FamiliasPedigree(id=persons, dadid=c(NA, 'AF', NA), momid=c(NA, 'mother', NA),
                         sex=c('female', 'female', 'male'))
datamatrix = data.frame(THO1.1=c(NA, 8, NA), THO1.2=c(NA,9.3, NA))
rownames(datamatrix) = persons
x = Familias2ped(ped1, datamatrix, locus1)
plotPedList(list(x), new=TRUE, frametitles=c('H1'), available='shaded', marker = 1)

# Example
library(fam2r)
data(F21)
pedigrees = F21$pedigrees
datamatrix = F21$datamatrix
loci = F21$loci
x = Familias2ped(pedigrees, datamatrix, loci)
plotPedList(x, new=TRUE, frametitles=c('H1', 'H2'), available='shaded')

# Give dev.width explicitly to allow for long names
plotPedList(x, new=TRUE, frametitles=c('H1', 'H2'), available='shaded',
            dev.width=17)

# Numerical labels work better
plotPedList(x, new=TRUE, id.labels='num', frametitles=c('H1', 'H2'),
            available='shaded')
}

}
\references{
Windows Familias is freely availabe from \url{http://familias.name}.
}
\author{
Magnus Dehli Vigeland, Thore Egeland
}
