% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forrel-package.R
\docType{package}
\name{forrel-package}
\alias{forrel}
\alias{forrel-package}
\title{forrel: Forensic Pedigree Analysis and Relatedness Inference}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Forensic applications of pedigree analysis, including likelihood ratios for relationship testing, general relatedness inference, marker simulation, and power analysis. 'forrel' is part of the 'pedsuite', a collection of packages for pedigree analysis, further described in the book 'Pedigree Analysis in R' (Vigeland, 2021, ISBN:9780128244302). Several functions deal specifically with power analysis in missing person cases, implementing methods described in Vigeland et al. (2020) \doi{10.1016/j.fsigen.2020.102376}. Data import from the 'Familias' software (Egeland et al. (2000) \doi{10.1016/S0379-0738(00)00147-X}) is supported through the 'pedFamilias' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/forrel}
  \item Report bugs at \url{https://github.com/magnusdv/forrel/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

Other contributors:
\itemize{
  \item Thore Egeland \email{thore.egeland@nmbu.no} [contributor]
}

}
\keyword{internal}
