% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipLR.R
\name{kinshipLR}
\alias{kinshipLR}
\title{Likelihood ratios for kinship testing}
\usage{
kinshipLR(..., ref = NULL, source = NULL, markers = NULL, verbose = FALSE)
}
\arguments{
\item{...}{A list of pedigree alternatives. Each alternative should be either
a single \code{ped} object or a list of such.}

\item{ref}{An index or name indicating which of the input pedigrees should be
used as "reference pedigree", i.e., used in the denominator of each LR. If
NULL (the default), the last pedigree is used as reference.}

\item{source}{An index or name designating one of the input pedigrees as
source for marker data. If given, marker data is transferred from this to
all the other pedigrees (replacing any existing markers). The default
action (\code{source = NULL}) is as follows: If all pedigrees already have
attached markers, no transfers are done. If some pedigrees are empty,
marker data is transferred to those, using the first nonempty pedigree as
source.}

\item{markers}{A vector of marker names or indices indicating which markers
should be included. If NULL (the default) all markers
are used.}

\item{verbose}{A logical.}
}
\value{
A \code{LRresult}object, which is essentially a list with entries
\itemize{
\item \code{LRtotal} : Total likelihood ratios
\item \code{LRperMarker} : Likelihood ratios for each marker
\item \code{likelihoodsPerMarker} : Likelihoods for each marker
\item \code{time} : Elapsed time
}
}
\description{
This function computes likelihood ratios (LRs) for a given a list of
pedigrees with attached markers. The user must indicate which of the
pedigrees is the 'reference', which will be used in the denominator in each
LR.
}
\examples{

# Simulate 5 markers for a pair of full sibs
ids = c("A", "B")
sibs = nuclearPed(children = ids)
sibs = simpleSim(sibs, N = 5, alleles = 1:4, ids = ids, seed = 123)

# Create two alternative hypotheses
halfsibs = relabel(halfSibPed(), old = 4:5, new = ids)
unrel = list(singleton("A"), singleton("B"))

# Compute LRs. By default, the last ped is used as reference
kinshipLR(sibs, halfsibs, unrel)

# Input pedigrees can be named, reflected in output
kinshipLR(S = sibs, H = halfsibs, U = unrel)

# Select non-default reference (by index or name)
kinshipLR(S = sibs, H = halfsibs, U = unrel, ref = "H")

# Alternative syntax: List input
peds = list(S = sibs, H = halfsibs, U = unrel)
kinshipLR(peds, ref = "H", source = "S", verbose = TRUE)

# Detailed results
res = kinshipLR(peds)
res$LRperMarker
res$likelihoodsPerMarker
}
\seealso{
\code{\link[=LRpower]{LRpower()}}, \code{\link[pedtools:transferMarkers]{pedtools::transferMarkers()}}
}
\author{
Magnus Dehli Vigeland and Thore Egeland
}
