% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonEP.R
\name{missingPersonEP}
\alias{missingPersonEP}
\title{Exclusion power for missing person cases}
\usage{
missingPersonEP(
  reference,
  missing,
  markers = NULL,
  disableMutations = NA,
  verbose = TRUE
)
}
\arguments{
\item{reference}{A \code{ped} object with attached markers.}

\item{missing}{The ID label of the missing pedigree member.}

\item{markers}{Marker input. Either:
\itemize{
\item \code{NULL} (default): Use all markers attached to the source pedigree.
\item A vector of names or indices selecting attached markers.
\item A named list of frequency vectors, replacing all existing marker data.
}

Ignored if \code{alleles} or \code{afreq} is given.}

\item{disableMutations}{This parameter determines how mutation models are
treated. Possible values are as follows:
\itemize{
\item \code{NA} (the default): Mutations are disabled only for those markers whose
known genotypes are consistent with the pedigree. This is determined by
temporarily removing all mutation models and checking which markers have
nonzero likelihood.
\item \code{TRUE}: Mutations are disabled for all markers. This will result in an
error if any markers are inconsistent.
\item \code{FALSE}: No action is done to disable mutations.
\item A vector containing the names or indices of those markers for which
mutations should be disabled.
}}

\item{verbose}{A logical.}
}
\value{
The \code{EPresult} object returned by \code{\link[=exclusionPower]{exclusionPower()}}.
}
\description{
This is a special case of \code{\link[=exclusionPower]{exclusionPower()}} for use in missing person cases.
The function computes the probability that a random person is genetically
incompatible with the typed relatives of the missing person.
}
\details{
This function is identical to \code{\link[=randomPersonEP]{randomPersonEP()}}, but with different argument
names. This makes it consistent with \code{\link[=missingPersonIP]{missingPersonIP()}} and the other
'missing person' functions.
}
\examples{

# Four siblings; the fourth is missing
x = nuclearPed(4)

# Remaining sibs typed with 4 triallelic markers
x = markerSim(x, N = 4, ids = 3:5, alleles = 1:3, seed = 577, verbose = FALSE)

# Add marker with inconsistency in reference genotypes
# (by default this is ignored by `missingPersonEP()`)
x = addMarker(x, "3" = "1/1", "4" = "2/2", "5" = "3/3")

# Compute exclusion power statistics
missingPersonEP(x, missing = 6)

}
\seealso{
\code{\link[=randomPersonEP]{randomPersonEP()}}, \code{\link[=exclusionPower]{exclusionPower()}}
}
