% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonEP.R
\name{missingPersonEP}
\alias{missingPersonEP}
\title{Exclusion power statistics in missing person cases}
\usage{
missingPersonEP(reference, missing, markers, disableMutations = TRUE,
  verbose = TRUE)
}
\arguments{
\item{reference}{A \code{\link[pedtools:ped]{pedtools::ped()}} object with attached markers.}

\item{missing}{The ID label of the missing pedigree member.}

\item{markers}{Names or indices of the markers to be included. By default, all markers are included.}

\item{disableMutations}{A logical (default: TRUE).}

\item{verbose}{A logical.}
}
\description{
Exclusion power statistics in missing person cases
}
\examples{

# Four siblings; the fourth is missing
x = nuclearPed(4)

# Remaining sibs typed with 4 triallelic markers
x = markerSim(x, N = 4, ids = 3:5, alleles = 1:3, seed = 577, verbose = FALSE)

# Add marker with inconsistency in reference genotypes
# (this should be ignored)
badMarker = marker(x, `3` = 1, `4` = 2, `5` = 3)
x = addMarkers(x, badMarker)

# Compute exclusion power statistics
missingPersonEP(x, missing = 6)

# Compare with genotypes
x

}
