% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonEP.R
\name{missingPersonEP}
\alias{missingPersonEP}
\title{Exclusion power statistics in missing person cases}
\usage{
missingPersonEP(reference, missing, markers, disableMutations = TRUE,
  verbose = TRUE)
}
\arguments{
\item{reference}{A \code{\link[pedtools:ped]{pedtools::ped()}} object with attached markers.}

\item{missing}{The ID label of the missing pedigree member.}

\item{markers}{Names or indices of the markers to be included. By default, all markers are included.}

\item{disableMutations}{A logical (default: TRUE).}

\item{verbose}{A logical.}
}
\description{
Exclusion power statistics in missing person cases
}
\examples{

# 3 siblings; third is missing
x = nuclearPed(3)

# First two brothers typed with 10 triallelic markers
x = markerSim(x, N = 5, ids = 3:4, alleles = 1:3, seed = 3, verbose = FALSE)
x

# Compute exclusion power statistics
missingPersonEP(x, missing = 5)

}
