% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonEP.R
\name{missingPersonEP}
\alias{missingPersonEP}
\title{Exclusion power for missing person cases.}
\usage{
missingPersonEP(reference, missing, markers = NULL,
  disableMutations = NA, verbose = TRUE)
}
\arguments{
\item{reference}{A \code{ped} object with attached markers.}

\item{missing}{The ID label of the missing pedigree member.}

\item{markers}{A vector indicating the names or indices of markers attached
to the source pedigree. (This is either \code{claimPed} or \code{truePed}, depending
on the value of \code{source}). If NULL (the default), then all markers attached
to the source pedigree are used. If \code{alleles} or \code{afreq} is non-NULL, then
this parameter is ignored.}

\item{disableMutations}{This parameter determines how mutation models are
treated:
\itemize{
\item If \code{NA} (the default), mutations are disabled only for those markers
whose known genotypes are compatible with both \code{claimPed} and \code{truePed}.
This is determined by temporarily removing all mutation models and checking
which markers have nonzero likelihood in both alternatives.
\item If \code{TRUE}, mutations are disabled for all markers.
\item If \code{FALSE} no action is done to disable mutations.
\item If a vector, it must be a subset of \code{markers} indicating explicitly those
markers for which mutations should be disabled.
}}

\item{verbose}{A logical.}
}
\value{
A \code{mpEP} object, which is essentially a list with the following entries:
\itemize{
\item \code{EPperMarker}: A numeric vector containing the exclusion power of each
marker. If the genotypes of a marker are incompatible with the \code{reference}
pedigree, the corresponding entry is NA
\item \code{EPtotal}: The total exclusion power, computed as \code{1 - prod(1 -   EPperMarker, na.rm = TRUE)}
\item \code{expectedMismatch}: The expected number of markers giving exclusion,
computed as \code{sum(EPperMarker, na.rm = TRUE)}
\item \code{distribMismatch}: The probability distribution of the number of markers
giving exclusion. This is given as a numeric vector of length \code{n+1}, where
\code{n} is the number of nonzero element of \code{EPperMarker}. The vector has names
\code{0:n}
\item \code{params}: A list containing the input parameters \code{missing}, \code{markers} and
\code{disableMutations}
}
}
\description{
Exclusion power for missing person cases.
}
\examples{

# Four siblings; the fourth is missing
x = nuclearPed(4)

# Remaining sibs typed with 4 triallelic markers
x = markerSim(x, N = 4, ids = 3:5, alleles = 1:3, seed = 577, verbose = FALSE)

# Add marker with inconsistency in reference genotypes
# (this should be ignored by `missingPersonEP()`)
badMarker = marker(x, `3` = 1, `4` = 2, `5` = 3)
x = addMarkers(x, badMarker)

# Compute exclusion power statistics
missingPersonEP(x, missing = 6)

# With marker names:
name(x, 1:5) = paste0("M", 1:5)
missingPersonEP(x, missing = 6)

}
