% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonIP.R
\name{missingPersonIP}
\alias{missingPersonIP}
\title{Inclusion power statistics in missing person cases.}
\usage{
missingPersonIP(reference, missing, markers, nsim = 1,
  threshold = NULL, disableMutations = NA, seed = NULL,
  verbose = T)
}
\arguments{
\item{reference}{A \code{ped} object with attached markers.}

\item{missing}{The ID label of the missing pedigree member.}

\item{markers}{Names or indices of the markers to be included. By default,
all markers.}

\item{nsim}{A positive integer: the number of simulations}

\item{threshold}{A numeric vector with one or more positive numbers used as
the likelihood ratio tresholds for inclusion}

\item{disableMutations}{This parameter determines how mutation models are
treated. If \code{TRUE}, mutations are disabled for all markers. If \code{NA} (the
default), mutations are disabled only for those markers with nonzero
baseline likelihood. (In other words: Mutations are NOT disabled if the
reference genotypes are inconsistent with the pedigree.) If \code{FALSE} no
action is done to disable mutations. Finally, if a vector of marker names
or indices is given, mutations are disabled for these markers exclusively.}

\item{seed}{A numeric seed for the random number generator (optional)}

\item{verbose}{A logical, by default TRUE.}
}
\value{
A \code{mpIP} object, which is essentially a list with the following
entries:
\itemize{
\item \code{LRperSim}: A numeric vector of length \code{nsim} containing the total LR for
each simulation.
\item \code{meanLRperMarker}: The mean LR per marker, over all simulations.
\item \code{meanLR}: The mean total LR over all simulations.
\item \code{meanLogLR}: The mean total \code{log10(LR)} over all simulations.
\item \code{IP}: A named numeric of the same length as \code{threshold}. For each element
of \code{threshold}, the fraction of simulations resulting in a LR exceeding the
given number.
\item \code{params}: A list containing the input parameters \code{missing}, \code{markers},
\code{nsim}, \code{threshold} and \code{disableMutations}
}
}
\description{
Inclusion power statistics in missing person cases.
}
\examples{

# Four siblings; the fourth is missing
x = nuclearPed(4)

# Remaining sibs typed with 5 triallelic markers
x = markerSim(x, N = 5, ids = 3:5, alleles = 1:3, seed = 123, verbose = FALSE)

# Compute exclusion power statistics
missingPersonIP(x, missing = 6, nsim = 5, threshold = c(10, 100))

# Compare with genotypes
x

}
