% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPersonPlot.R
\name{missingPersonPlot}
\alias{missingPersonPlot}
\title{Missing person plot}
\usage{
missingPersonPlot(
  reference,
  missing,
  id.labels = labels(reference),
  MP.label = "MP",
  POI.label = "POI",
  marker = NULL,
  shaded = "typed",
  POI.col = "red",
  POI.shaded = FALSE,
  POI.height = 8,
  width = 4,
  newdev = TRUE,
  frametitles = c(expression(H[1] * ": POI is MP"), expression(H[2] *
    ": POI unrelated")),
  ...
)
}
\arguments{
\item{reference}{A \code{\link[pedtools:ped]{pedtools::ped()}} object.}

\item{missing}{The ID label of the missing pedigree member.}

\item{id.labels}{A character vector with labels for the pedigree members. See
\code{\link[pedtools:plot.ped]{pedtools::plot.ped()}}.}

\item{MP.label}{The label of the missing member. Default: "MP".}

\item{POI.label}{The label of the person of interest. Default: "POI".}

\item{marker}{Optional vector of marker indices to be included in the plot.}

\item{shaded}{A vector of ID labels indicating who should appear with shaded
symbols. By default, all typed members.}

\item{POI.col}{The plot colour of POI. Default: red.}

\item{POI.shaded}{A logical: If TRUE (default), the POI is plotted with a
shaded symbol.}

\item{POI.height}{A numeric controlling the vertical placement of the POI
singleton (in the right panel).}

\item{width}{A positive number controlling the width of the plot. More
specifically this number is the relative width of the reference pedigree,
compared to a singleton. Default: 4.}

\item{newdev}{A logical: If TRUE the plot is created in a new plot window.}

\item{frametitles}{A character of length 2, with subtitles for the two frames.}

\item{...}{Extra parameters passed on to \code{\link[pedtools:plotPedList]{pedtools::plotPedList()}}.}
}
\description{
Visualises the competing hypotheses of a family reunion case. A plot with two
panels is generated. The left panel shows a pedigree in which the \emph{person of
interest} (POI) is identical to the \emph{missing person} (MP). The right panel
shows the situation where these two are unrelated. See Details for further
explanations.
}
\details{
A standard family reunification case involves the following ingredients:
\itemize{
\item A reference family in which a single member ("MP") is missing.
\item Some of the family members have been genotyped
\item A person of interest ("POI") is to be matched against the reference family
}

After genotyping of POI, the genetic evidence is typically assessed by
computing the likelihood ratio of the following hypotheses:
\itemize{
\item H1: POI is MP
\item H2: POI is unrelated to the family
}

The goal of this function is to illustrate the above hypotheses, using
labels, colours and shading to visualise the different aspects of the
situation.

This function cannot handle cases with more complicated hypotheses (e.g.
multiple missing persons, or where H2 specifies a different relationship).
However, as it is basically a wrapper of \code{\link[pedtools:plotPedList]{pedtools::plotPedList()}}, an
interested user should be able to extend the source code to such cases
without too much trouble.
}
\examples{
x = nuclearPed(father = "fa", mother = "mo", children = c("b1", "b2"))

# Default plot
missingPersonPlot(x, missing = "b2")

# A bit nicer using various options
missingPersonPlot(x, missing = "b2", MP.label = "Missing", id.label = NULL,
                  shaded = "b1", POI.shaded = TRUE,
                  width = 2,      # adjust internal spacing (see above)
                  dev.width = 7,  # device width (see ?plotPedList())
                  dev.height = 3, # device height (see ?plotPedList())
                  fmar = 0.02,    # adjust frame margin (see ?plotPedList())
                  cex = 1.5,      # larger symbols and label font (see ?par())
                  cex.main = 1.3  # larger frame titles (see ?par())
                  )

}
