% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPsims.R
\name{MPPsims}
\alias{MPPsims}
\title{Missing person power simulations}
\usage{
MPPsims(
  reference,
  missing = "MP",
  selections,
  baseline = TRUE,
  nProfiles = 1,
  lrSims = 1,
  thresholdIP = NULL,
  verbose = TRUE
)
}
\arguments{
\item{reference}{A connected \code{ped} object}

\item{missing}{The label of the missing person. By default "MP".}

\item{selections}{A list of pedigree member subsets. In the special case that
all subsets consist of a single individual, \code{selections} can be given as a
simple vector.}

\item{baseline}{A logical. If TRUE (default) an \emph{empty} selection, named
"Baseline", is added as the first element of \code{selection}.}

\item{nProfiles}{The number of profile simulations for each selection.}

\item{lrSims, thresholdIP}{Parameters passed onto \code{\link[=missingPersonIP]{missingPersonIP()}}}

\item{verbose}{A logical}
}
\value{
A list with the same length as \code{selections}. Each entry has elements
\code{ep} and \code{ip}, each of which is a list of length \code{nProfiles}.
}
\description{
Estimate the exclusion/inclusion power for various selections of available
individuals.
}
\examples{
x = nuclearPed(fa = "Gf", mo = "Gm", children = c("Uncle", "Mother"), sex = 1:2)
x = addChildren(x, fa = "Father", mo = "Mother", nch = 3, sex = c(1,2,1),
                id = c("S1", "S2", "MP"))
x = addSon(x, "Father", id = "HS")

# Brother S1 is already genotyped with a marker with 5 alleles
m = marker(x, S1 = 1:2, alleles = 1:5)
x = setMarkers(x, m)

# Alternatives for additional genotyping
sel = list("Father", "S2", "HS", "Gm", "Uncle", c("Gm", "Uncle"))

plot(x, marker = 1, shaded = sel)

# Simulate
simData = MPPsims(x, selections = sel, nProfiles = 10, lrSims = 10)

# Power plot
epSims = lapply(simData, '[[', 'ep')
ipSims = lapply(simData, '[[', 'ip')
powerPlot(epSims, ipSims, type = 3)

}
