% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFam.R
\name{readFam}
\alias{readFam}
\title{Read Familias .fam files}
\usage{
readFam(famfile)
}
\arguments{
\item{famfile}{Path to a ".fam" file}
}
\value{
If the .fam file only contains a database, the output is a list of
information (name, alleles, frequencies) about each locus. This list can be
used as \code{locusAttributes} in e.g. \code{\link[=setMarkers]{setMarkers()}}.

If the .fam file describes pedigree data, the output is a \code{ped} object or a
list of such.
}
\description{
This function parses the content of a Familias-formatted ".fam" file, and
converts it into suitable \code{ped} objects. This function does not depend on the
\code{Familias} package.
}
