% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFam.R
\name{readFam}
\alias{readFam}
\title{Read Familias .fam files}
\usage{
readFam(famfile, useDVI = NA, verbose = T)
}
\arguments{
\item{famfile}{Path to a ".fam" file.}

\item{useDVI}{A logical, or NA. If TRUE, the DVI section of the fam file is
used to extract pedigrees and genotypes. If NA (default), the parameter is
set to TRUE if the input data contains a line equal to "\link{DVI}", and
otherwise FALSE.}

\item{verbose}{A logical; if TRUE, various information is written to the
screen during the parsing process.}
}
\value{
If the .fam file only contains a database, the output is a list of
information (name, alleles, frequencies) about each locus. This list can be
used as \code{locusAttributes} in e.g. \code{\link[=setMarkers]{setMarkers()}}.

If the .fam file describes pedigree data, the output is a \code{ped} object or a
list of such.

If \code{useDVI = TRUE}, then the families described under \code{Reference Families}
are parsed and converted to \code{ped} objects. Each family generally describes
multiple pedigrees, so the output gets another layer in this case.
}
\description{
This function parses the content of a Familias-formatted ".fam" file, and
converts it into suitable \code{ped} objects. This function does not depend on the
\code{Familias} package.
}
