% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showInTriangle}
\alias{showInTriangle}
\title{Add points to the IBD triangle}
\description{
This function is re-exported from the \code{ribd} package. For documentation see
\code{\link[ribd:showInTriangle]{ribd::showInTriangle()}}.
}
